feat: add unit tests for Plot1D and PlotBar functionality
